requirements.txt
PyMuPDF
transformers
torch
pymongo
yake
concurrent-log-handler
cohere
python-dotenv